+++ NEWS.md
@@ -1,3 +1,6 @@
+* Added efficiency in engineering units (e.g., mpg) to stages table.
+
+
 # ReboundTools 0.1.8 (2021-01-16)
 
 * Fixed a bug where `Re_prod` was counted as a subtotal.